Move css files to styles folder
