+++ src/components/InputBook.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { useDispatch } from 'react-redux';
 import { v4 as uuidv4 } from 'uuid';
 import { addBook } from '../redux/books/books';
-import './InputBook.css';
+import '../styles/InputBook.css';
 
 const InputBook = () => {
   const dispatch = useDispatch();
