+++ src/styles/App.css
@@ -4,11 +4,10 @@
   --azure: #0290ff;
   --black-two: #121212;
   --pale-gray: #9c9c9c;
-  --border: rgb(212, 208, 208);
-  --btn: #4386bf;
   --warm-gray: #5a5959;
   --white: #e8e8e8;
   --pinkish-grey: #c4c4c4;
+  --dodger-blue: #379cf6;
 }
 
 body {
